requirements.txt
geopandas
pyogrio
shapely
fiona
rasterio
pandas
numpy
tqdm
pyyaml
earthengine-api
geemap
python-dotenv
duckdb